# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Ruth Health Products & Services API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  imageUrl: string | null;
  guestPrice: number;
  memberPrice: number;
  commissionPct: number;
  createdAt: Date;
}